Flatten request.services namespace, drop chronology/panel scaffolding
Same reasoning as the repository registry flattening: the
chronology.panel.* nesting was speculative scaffold around a single leaf.
Collapse to `request.services.personal_data_fields` and let buckets
appear when the leaf count grows past ~12.
