+++ src/ludamus/pacts/services.py
@@ -1,8 +1,8 @@
 """Service-side infrastructure and navigation protocols.
 
 Holds the cross-cutting protocols that describe how mills services are wired
-and reached from gates: the transaction adapter and the lazy nested
-`request.services` namespace tree.
+and reached from gates: the transaction adapter and the flat
+`request.services` namespace.
 """
 
 from typing import TYPE_CHECKING, Protocol
@@ -18,16 +18,6 @@ class TransactionProtocol(Protocol):
     def atomic() -> AbstractContextManager[None]: ...
 
 
-class ChronologyPanelServicesProtocol(Protocol):
-    @property
-    def personal_data_fields(self) -> CFPPersonalDataFieldServiceProtocol: ...
-
-
-class ChronologyServicesProtocol(Protocol):
-    @property
-    def panel(self) -> ChronologyPanelServicesProtocol: ...
-
-
 class ServicesProtocol(Protocol):
     @property
-    def chronology(self) -> ChronologyServicesProtocol: ...
+    def personal_data_fields(self) -> CFPPersonalDataFieldServiceProtocol: ...
